DEVPROD-19304: Add pagination buttons to bottom of patches page
