+++ apps/spruce/src/components/PatchesPage/index.tsx
@@ -1,12 +1,9 @@
 import styled from "@emotion/styled";
 import Checkbox from "@leafygreen-ui/checkbox";
 import Cookies from "js-cookie";
-import PageSizeSelector from "@evg-ui/lib/components/PageSizeSelector";
-import Pagination from "@evg-ui/lib/components/Pagination";
 import { size } from "@evg-ui/lib/constants/tokens";
 import { useQueryParam } from "@evg-ui/lib/hooks";
 import { usePageTitle } from "@evg-ui/lib/hooks/usePageTitle";
-import usePagination from "@evg-ui/lib/src/hooks/usePagination";
 import { useProjectPatchesAnalytics, useUserPatchesAnalytics } from "analytics";
 import { PageWrapper, FiltersWrapper, PageTitle } from "components/styles";
 import TextInputWithValidation from "components/TextInputWithValidation";
@@ -16,8 +13,8 @@ import { useFilterInputChangeHandler } from "hooks";
 import { PatchPageQueryParams } from "types/patch";
 import { validateRegexp } from "utils/validators";
 import ListArea from "./ListArea";
+import { PaginationButtons } from "./PaginationButtons";
 import { StatusSelector } from "./StatusSelector";
-import { usePatchesQueryParams } from "./usePatchesQueryParams";
 
 interface Props {
   filterComp?: React.ReactNode;
@@ -39,14 +36,6 @@ export const PatchesPage: React.FC<Props> = ({
   const projectPatchesAnalytics = useProjectPatchesAnalytics();
   const analytics =
     pageType === "project" ? projectPatchesAnalytics : userPatchesAnalytics;
-
-  // Handle pagination logic.
-  const { setLimit } = usePagination();
-  const { limit, page } = usePatchesQueryParams();
-  const handlePageSizeChange = (pageSize: number): void => {
-    setLimit(pageSize);
-    analytics.sendEvent({ name: "Changed page size" });
-  };
 
   // Handle filtering by patch description.
   const { setAndSubmitInputValue } = useFilterInputChangeHandler({
@@ -97,32 +86,24 @@ export const PatchesPage: React.FC<Props> = ({
           onChange={includeHiddenCheckboxOnChange}
         />
       </FiltersWrapperSpaceBetween>
-      <PaginationRow>
-        <Pagination
-          currentPage={page}
-          pageSize={limit}
-          totalResults={patches?.filteredPatchCount ?? 0}
-        />
-        <PageSizeSelector
-          data-cy="my-patches-page-size-selector"
-          onChange={handlePageSizeChange}
-          value={limit}
-        />
-      </PaginationRow>
+      <PaginationButtons
+        filteredPatchCount={patches?.filteredPatchCount}
+        pageType={pageType}
+      />
       <ListArea
         loading={loading}
         pageType={pageType}
         patches={patches?.patches || []}
       />
+      {!loading && (
+        <PaginationButtons
+          filteredPatchCount={patches?.filteredPatchCount}
+          pageType={pageType}
+        />
+      )}
     </PageWrapper>
   );
 };
-
-const PaginationRow = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-`;
 
 const FiltersWrapperSpaceBetween = styled(FiltersWrapper)`
   display: grid;
